+++ .gitignore
@@ -3,7 +3,6 @@ typings
 node_modules
 coverage
 doc
-dist/*.js
-dist/*.map
+dist
 *.log
 .idea/
