+++ package.json
@@ -10,72 +10,53 @@
   "author": "js-data-adapter project authors",
   "license": "MIT",
   "main": "./dist/js-data-adapter.js",
-  "typings": "./dist/js-data-adapter.d.ts",
+  "typings": "./dist/index.d.ts",
   "files": [
     "dist/",
-    "src/",
-    "test/",
     "AUTHORS",
-    "CONTRIBUTORS",
-    "typings.json"
+    "CONTRIBUTORS"
   ],
   "keywords": [
     "js-data",
     "jsdata",
     "adapter"
   ],
-  "standard": {
-    "parser": "babel-eslint",
-    "globals": [
-      "Mocha",
-      "describe",
-      "sinon",
-      "it",
-      "before",
-      "after",
-      "beforeEach",
-      "afterEach",
-      "assert"
-    ],
-    "ignore": [
-      "dist/"
-    ]
-  },
   "scripts": {
-    "lint": "standard lint '**/*.js'",
-    "bundle_adapter": "rollup src/index.js -c -o dist/js-data-adapter.js -m dist/js-data-adapter.js.map -f umd",
-    "bundle_tests": "rollup test/index.js -c rollup-tests.config.js -o dist/js-data-adapter-tests.js -m dist/js-data-adapter-tests.js.map -f umd",
-    "bundle": "npm run bundle_adapter && npm run bundle_tests",
-    "watch": "watch \"npm run bundle\" src/",
+    "lint": "eslint --fix '{src,test}/**/*.ts'",
+    "bundle": "rollup src/index.ts -c -o dist/js-data-adapter.js -m dist/js-data-adapter.js.map -f umd",
     "build": "npm run lint && npm run bundle",
-    "mocha": "mocha -t 2000 -R dot mocha.start.js -r babel-core/register -r babel-polyfill",
-    "cover": "nyc --require babel-core/register --require babel-polyfill --cache mocha mocha.start.js -t 2000 -R dot && nyc report --reporter=html",
+    "mocha": "mocha -t 2000 -R dot './test/**/*.test.ts' -r ts-node/register",
+    "cover": "nyc --require ts-node/register --cache mocha './test/**/*.test.ts' -t 2000 -R dot && nyc report --reporter=html",
     "test": "npm run build && npm run cover",
+    "doc": "typedoc --mode node --out doc src",
     "release": "npm test && repo-tools updates && repo-tools changelog && repo-tools authors"
   },
-  "peerDependencies": {
-    "js-data": ">= 3.0.0"
-  },
   "dependencies": {
-    "js-data": ">= 3.0.0"
+    "js-data": "^4.0.0-beta.3"
   },
   "devDependencies": {
-    "babel-core": "6.25.0",
-    "babel-eslint": "7.2.3",
-    "babel-plugin-external-helpers": "6.22.0",
-    "babel-plugin-syntax-async-functions": "6.13.0",
-    "babel-plugin-transform-es2015-modules-umd": "6.24.1",
-    "babel-plugin-transform-regenerator": "6.24.1",
-    "babel-polyfill": "6.23.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-stage-0": "6.24.1",
-    "js-data-repo-tools": "1.0.0",
-    "nyc": "11.0.3",
-    "rollup": "0.43.0",
-    "rollup-plugin-babel": "2.7.1",
-    "standard": "10.0.2",
-    "sinon": "^2.3.6",
-    "chai": "^4.0.2",
-    "mocha": "^3.4.2"
+    "@js-data/repo-tools": "^1.0.1",
+    "@types/chai": "^4.2.7",
+    "@types/mocha": "^5.2.7",
+    "@types/sinon": "^7.5.1",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "chai": "^4.2.0",
+    "eslint": "^6.8.0",
+    "eslint-config-standard-with-typescript": "^12.0.1",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.20.0",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "mocha": "^7.0.1",
+    "nyc": "^15.0.0",
+    "rollup": "^1.29.0",
+    "rollup-plugin-typescript2": "^0.25.3",
+    "sinon": "^8.1.0",
+    "standard": "^14.3.1",
+    "ts-node": "^8.6.2",
+    "tslib": "^1.10.0",
+    "typedoc": "^0.16.9",
+    "typescript": "^3.7.5"
   }
 }
