+++ src/index.ts
@@ -1,21 +1,19 @@
-import { Component, utils } from 'js-data'
+import { Component, utils, Mapper } from 'js-data'
 
-export const noop = function (...args) {
+export async function noop (...args) {
   const opts = args[args.length - 1]
   this.dbg(opts.op, ...args)
-  return utils.resolve()
 }
 
-export const noop2 = function (...args) {
+export async function noop2 (...args) {
   const opts = args[args.length - 2]
   this.dbg(opts.op, ...args)
-  return utils.resolve()
 }
 
-export const unique = function (array) {
+export const unique = array => {
   const seen = {}
   const final = []
-  array.forEach(function (item) {
+  array.forEach(item => {
     if (item in seen) {
       return
     }
@@ -25,7 +23,7 @@ export const unique = function (array) {
   return final
 }
 
-export const withoutRelations = function (mapper, props, opts) {
+export const withoutRelations = (mapper, props, opts) => {
   opts || (opts = {})
   opts.with || (opts.with = [])
   const relationFields = mapper.relationFields || []
@@ -48,26 +46,28 @@ export const reserved = [
  *
  * @class Response
  */
-export function Response (data, meta, op) {
-  meta || (meta = {})
+export class Response {
+  created: number
+  found: number
+  updated: number
 
-  /**
-   * Response data.
-   *
-   * @name Response#data
-   * @type {*}
-   */
-  this.data = data
-
-  utils.fillIn(this, meta)
-
-  /**
-   * The operation for which the response was created.
-   *
-   * @name Response#op
-   * @type {string}
-   */
-  this.op = op
+  constructor (
+    /**
+     * Response data.
+     *
+     * @name Response#data
+     * @type {*}
+     */
+    public data,
+    meta = {},
+    /**
+     * The operation for which the response was created.
+     *
+     * @type {string}
+     */
+    public op: string) {
+    utils.fillIn(this, meta)
+  }
 }
 
 const DEFAULTS = {
@@ -93,24 +93,28 @@ const DEFAULTS = {
 /**
  * Abstract class meant to be extended by adapters.
  *
- * @class Adapter
+ * @example
+ * class CustomAdapterClass extends Adapter {
+ *   foo () { return 'bar' }
+ *   static beep () { return 'boop' }
+ * }
+ * const customAdapter = new CustomAdapterClass()
+ * console.log(customAdapter.foo())
+ * console.log(CustomAdapterClass.beep())
  * @abstract
- * @extends Component
- * @param {Object} [opts] Configuration opts.
- * @param {boolean} [opts.debug=false] Whether to log debugging information.
- * @param {boolean} [opts.raw=false] Whether to return a more detailed response
- * object.
  */
-export function Adapter (opts) {
-  utils.classCallCheck(this, Adapter)
-  Component.call(this, opts)
-  opts || (opts = {})
-  utils.fillIn(opts, DEFAULTS)
-  utils.fillIn(this, opts)
-}
-
-Component.extend({
-  constructor: Adapter,
+export abstract class Adapter extends Component {
+  /**
+   * @param {Object} [opts] Configuration opts.
+   * @param {boolean} [opts.debug=false] Whether to log debugging information.
+   * @param {boolean} [opts.raw=false] Whether to return a more detailed response
+   * object.
+   */
+  constructor (opts = {}) {
+    super(opts)
+    utils.fillIn(opts, DEFAULTS)
+    utils.fillIn(this, opts)
+  }
 
   /**
    * Lifecycle method method called by <a href="#count__anchor">count</a>.
@@ -133,7 +137,7 @@ Component.extend({
    * @property {string} opts.op `afterCount`
    * @param {Object|Response} response Count or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterCount: noop2,
+  afterCount = noop2
 
   /**
    * Lifecycle method method called by <a href="#create__anchor">create</a>.
@@ -156,7 +160,7 @@ Component.extend({
    * @property {string} opts.op `afterCreate`
    * @param {Object|Response} response Created record or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterCreate: noop2,
+  afterCreate = noop2
 
   /**
    * Lifecycle method method called by <a href="#createMany__anchor">createMany</a>.
@@ -179,7 +183,7 @@ Component.extend({
    * @property {string} opts.op `afterCreateMany`
    * @param {Object[]|Response} response Created records or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterCreateMany: noop2,
+  afterCreateMany = noop2
 
   /**
    * Lifecycle method method called by <a href="#destroy__anchor">destroy</a>.
@@ -202,7 +206,7 @@ Component.extend({
    * @property {string} opts.op `afterDestroy`
    * @param {undefined|Response} response `undefined` or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterDestroy: noop2,
+  afterDestroy = noop2
 
   /**
    * Lifecycle method method called by <a href="#destroyAll__anchor">destroyAll</a>.
@@ -225,7 +229,7 @@ Component.extend({
    * @property {string} opts.op `afterDestroyAll`
    * @param {undefined|Response} response `undefined` or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterDestroyAll: noop2,
+  afterDestroyAll = noop2
 
   /**
    * Lifecycle method method called by <a href="#find__anchor">find</a>.
@@ -248,7 +252,9 @@ Component.extend({
    * @property {string} opts.op `afterFind`
    * @param {Object|Response} response The found record or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterFind: noop2,
+  afterFind (mapper, id, opts, response) {
+    return noop2.call(this, mapper, id, opts)
+  }
 
   /**
    * Lifecycle method method called by <a href="#findAll__anchor">findAll</a>.
@@ -271,7 +277,7 @@ Component.extend({
    * @property {string} opts.op `afterFindAll`
    * @param {Object[]|Response} response The found records or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterFindAll: noop2,
+  afterFindAll = noop2
 
   /**
    * Lifecycle method method called by <a href="#sum__anchor">sum</a>.
@@ -295,7 +301,7 @@ Component.extend({
    * @property {string} opts.op `afterSum`
    * @param {Object|Response} response Count or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterSum: noop2,
+  afterSum = noop2
 
   /**
    * Lifecycle method method called by <a href="#update__anchor">update</a>.
@@ -319,7 +325,7 @@ Component.extend({
    * @property {string} opts.op `afterUpdate`
    * @param {Object|Response} response The updated record or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterUpdate: noop2,
+  afterUpdate = noop2
 
   /**
    * Lifecycle method method called by <a href="#updateAll__anchor">updateAll</a>.
@@ -343,7 +349,7 @@ Component.extend({
    * @property {string} opts.op `afterUpdateAll`
    * @param {Object[]|Response} response The updated records or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterUpdateAll: noop2,
+  afterUpdateAll = noop2
 
   /**
    * Lifecycle method method called by <a href="#updateMany__anchor">updateMany</a>.
@@ -366,7 +372,7 @@ Component.extend({
    * @property {string} opts.op `afterUpdateMany`
    * @param {Object[]|Response} response The updated records or {@link Response}, depending on the value of `opts.raw`.
    */
-  afterUpdateMany: noop2,
+  afterUpdateMany = noop2
 
   /**
    * Lifecycle method method called by <a href="#count__anchor">count</a>.
@@ -384,7 +390,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#count__anchor">count</a>.
    * @property {string} opts.op `beforeCount`
    */
-  beforeCount: noop,
+  beforeCount = noop
 
   /**
    * Lifecycle method method called by <a href="#create__anchor">create</a>.
@@ -404,7 +410,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#create__anchor">create</a>.
    * @property {string} opts.op `beforeCreate`
    */
-  beforeCreate: noop,
+  beforeCreate = noop
 
   /**
    * Lifecycle method method called by <a href="#createMany__anchor">createMany</a>.
@@ -424,7 +430,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#createMany__anchor">createMany</a>.
    * @property {string} opts.op `beforeCreateMany`
    */
-  beforeCreateMany: noop,
+  beforeCreateMany = noop
 
   /**
    * Lifecycle method method called by <a href="#destroy__anchor">destroy</a>.
@@ -442,7 +448,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#destroy__anchor">destroy</a>.
    * @property {string} opts.op `beforeDestroy`
    */
-  beforeDestroy: noop,
+  beforeDestroy = noop
 
   /**
    * Lifecycle method method called by <a href="#destroyAll__anchor">destroyAll</a>.
@@ -460,7 +466,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#destroyAll__anchor">destroyAll</a>.
    * @property {string} opts.op `beforeDestroyAll`
    */
-  beforeDestroyAll: noop,
+  beforeDestroyAll = noop
 
   /**
    * Lifecycle method method called by <a href="#find__anchor">find</a>.
@@ -478,7 +484,9 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#find__anchor">find</a>.
    * @property {string} opts.op `beforeFind`
    */
-  beforeFind: noop,
+  beforeFind (mapper, id, opts) {
+    return noop.call(this, mapper, id, opts)
+  }
 
   /**
    * Lifecycle method method called by <a href="#findAll__anchor">findAll</a>.
@@ -496,7 +504,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#findAll__anchor">findAll</a>.
    * @property {string} opts.op `beforeFindAll`
    */
-  beforeFindAll: noop,
+  beforeFindAll = noop
 
   /**
    * Lifecycle method method called by <a href="#sum__anchor">sum</a>.
@@ -514,7 +522,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#sum__anchor">sum</a>.
    * @property {string} opts.op `beforeSum`
    */
-  beforeSum: noop,
+  beforeSum = noop
 
   /**
    * Lifecycle method method called by <a href="#update__anchor">update</a>.
@@ -535,7 +543,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#update__anchor">update</a>.
    * @property {string} opts.op `beforeUpdate`
    */
-  beforeUpdate: noop,
+  beforeUpdate = noop
 
   /**
    * Lifecycle method method called by <a href="#updateAll__anchor">updateAll</a>.
@@ -556,7 +564,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#updateAll__anchor">updateAll</a>.
    * @property {string} opts.op `beforeUpdateAll`
    */
-  beforeUpdateAll: noop,
+  beforeUpdateAll = noop
 
   /**
    * Lifecycle method method called by <a href="#updateMany__anchor">updateMany</a>.
@@ -576,7 +584,7 @@ Component.extend({
    * @param {Object} opts The `opts` argument passed to <a href="#updateMany__anchor">updateMany</a>.
    * @property {string} opts.op `beforeUpdateMany`
    */
-  beforeUpdateMany: noop,
+  beforeUpdateMany = noop
 
   /**
    * Retrieve the number of records that match the selection query. Called by
@@ -604,7 +612,7 @@ Component.extend({
 
     // beforeCount lifecycle hook
     op = opts.op = 'beforeCount'
-    return utils.resolve(this[op](mapper, query, opts))
+    return utils.resolve(this.beforeCount(mapper, query, opts))
       .then(() => {
         // Allow for re-assignment from lifecycle hook
         op = opts.op = 'count'
@@ -619,10 +627,12 @@ Component.extend({
 
         // afterCount lifecycle hook
         op = opts.op = 'afterCount'
-        return utils.resolve(this[op](mapper, query, opts, response))
+        return utils.resolve(this.afterCount(mapper, query, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _count (mapper: any, query: any, opts: any): any;
 
   /**
    * Create a new record. Called by `Mapper#create`.
@@ -636,14 +646,12 @@ Component.extend({
    * response object.
    * @return {Promise}
    */
-  create (mapper, props, opts) {
+  create (mapper, props = {}, opts: any = {}) {
     let op
-    props || (props = {})
-    opts || (opts = {})
 
     // beforeCreate lifecycle hook
     op = opts.op = 'beforeCreate'
-    return utils.resolve(this[op](mapper, props, opts))
+    return utils.resolve(this.beforeCreate(mapper, props, opts))
       .then((_props) => {
         // Allow for re-assignment from lifecycle hook
         props = _props === undefined ? props : _props
@@ -661,10 +669,12 @@ Component.extend({
 
         // afterCreate lifecycle hook
         op = opts.op = 'afterCreate'
-        return utils.resolve(this[op](mapper, props, opts, response))
+        return utils.resolve(this.afterCreate(mapper, props, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _create (mapper: any, props: any, opts: any): any
 
   /**
    * Create multiple records in a single batch. Called by `Mapper#createMany`.
@@ -685,7 +695,7 @@ Component.extend({
 
     // beforeCreateMany lifecycle hook
     op = opts.op = 'beforeCreateMany'
-    return utils.resolve(this[op](mapper, props, opts))
+    return utils.resolve(this.beforeCreateMany(mapper, props, opts))
       .then((_props) => {
         // Allow for re-assignment from lifecycle hook
         props = _props === undefined ? props : _props
@@ -704,10 +714,12 @@ Component.extend({
 
         // afterCreateMany lifecycle hook
         op = opts.op = 'afterCreateMany'
-        return utils.resolve(this[op](mapper, props, opts, response))
+        return utils.resolve(this.afterCreateMany(mapper, props, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _createMany (mapper: Mapper, props: any, opts: any): any
 
   /**
    * Destroy the record with the given primary key. Called by
@@ -728,7 +740,7 @@ Component.extend({
 
     // beforeDestroy lifecycle hook
     op = opts.op = 'beforeDestroy'
-    return utils.resolve(this[op](mapper, id, opts))
+    return utils.resolve(this.beforeDestroy(mapper, id, opts))
       .then(() => {
         op = opts.op = 'destroy'
         this.dbg(op, mapper, id, opts)
@@ -742,10 +754,12 @@ Component.extend({
 
         // afterDestroy lifecycle hook
         op = opts.op = 'afterDestroy'
-        return utils.resolve(this[op](mapper, id, opts, response))
+        return utils.resolve(this.afterDestroy(mapper, id, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _destroy (mapper: any, id: any, opts: any): any
 
   /**
    * Destroy the records that match the selection query. Called by
@@ -773,7 +787,7 @@ Component.extend({
 
     // beforeDestroyAll lifecycle hook
     op = opts.op = 'beforeDestroyAll'
-    return utils.resolve(this[op](mapper, query, opts))
+    return utils.resolve(this.beforeDestroyAll(mapper, query, opts))
       .then(() => {
         op = opts.op = 'destroyAll'
         this.dbg(op, mapper, query, opts)
@@ -787,10 +801,12 @@ Component.extend({
 
         // afterDestroyAll lifecycle hook
         op = opts.op = 'afterDestroyAll'
-        return utils.resolve(this[op](mapper, query, opts, response))
+        return utils.resolve(this.afterDestroyAll(mapper, query, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _destroyAll (mapper: any, query: any, opts: any): any
 
   /**
    * Load a belongsTo relationship.
@@ -817,7 +833,7 @@ Component.extend({
       return this.findAll(relationDef, {
         where: {
           [relationDef.idAttribute]: {
-            'in': keys
+            in: keys
           }
         }
       }, __opts).then((relatedItems) => {
@@ -830,7 +846,7 @@ Component.extend({
         })
       })
     }
-  },
+  }
 
   /**
    * Retrieve the record with the given primary key. Called by `Mapper#find`.
@@ -852,7 +868,7 @@ Component.extend({
 
     // beforeFind lifecycle hook
     op = opts.op = 'beforeFind'
-    return utils.resolve(this[op](mapper, id, opts))
+    return utils.resolve(this.beforeFind(mapper, id, opts))
       .then(() => {
         op = opts.op = 'find'
         this.dbg(op, mapper, id, opts)
@@ -866,10 +882,12 @@ Component.extend({
 
         // afterFind lifecycle hook
         op = opts.op = 'afterFind'
-        return utils.resolve(this[op](mapper, id, opts, response))
+        return utils.resolve(this.afterFind(mapper, id, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _find (mapper: any, id: any, opts: any): any
 
   /**
    * Retrieve the records that match the selection query.
@@ -908,7 +926,7 @@ Component.extend({
 
     // beforeFindAll lifecycle hook
     op = opts.op = 'beforeFindAll'
-    return utils.resolve(this[op](mapper, query, opts))
+    return utils.resolve(this.beforeFindAll(mapper, query, opts))
       .then(() => {
         op = opts.op = 'findAll'
         this.dbg(op, mapper, query, opts)
@@ -922,10 +940,12 @@ Component.extend({
 
         // afterFindAll lifecycle hook
         op = opts.op = 'afterFindAll'
-        return utils.resolve(this[op](mapper, query, opts, response))
+        return utils.resolve(this.afterFindAll(mapper, query, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
+
+  abstract _findAll (mapper: any, query: any, opts: any): any
 
   loadRelationsFor (mapper, results, opts) {
     const [records] = results
@@ -953,9 +973,9 @@ Component.extend({
       })
     }
 
-    return utils.Promise.all(tasks)
+    return Promise.all(tasks)
       .then(() => results)
-  },
+  }
 
   /**
    * Resolve the value of the specified option based on the given options and
@@ -970,7 +990,7 @@ Component.extend({
   getOpt (opt, opts) {
     opts || (opts = {})
     return opts[opt] === undefined ? utils.plainCopy(this[opt]) : utils.plainCopy(opts[opt])
-  },
+  }
 
   /**
    * Load a hasMany relationship.
@@ -992,12 +1012,12 @@ Component.extend({
     const query = {
       where: {}
     }
-    const criteria = query.where[def.foreignKey] = {}
+    const criteria: any = query.where[def.foreignKey] = {}
     if (singular) {
       // more efficient query when we only have one record
       criteria['=='] = IDs[0]
     } else {
-      criteria['in'] = IDs.filter((id) => id)
+      criteria.in = IDs.filter((id) => id)
     }
     return this.findAll(def.getRelation(), query, __opts).then((relatedItems) => {
       records.forEach((record) => {
@@ -1015,7 +1035,7 @@ Component.extend({
         def.setLocalField(record, attached)
       })
     })
-  },
+  }
 
   loadHasManyLocalKeys (mapper, def, records, __opts) {
     let record
@@ -1029,7 +1049,7 @@ Component.extend({
       return this.findAll(relatedMapper, {
         where: {
           [relatedMapper.idAttribute]: {
-            'in': this.makeHasManyLocalKeys(mapper, def, record)
+            in: this.makeHasManyLocalKeys(mapper, def, record)
           }
         }
       }, __opts).then((relatedItems) => {
@@ -1043,12 +1063,12 @@ Component.extend({
       return this.findAll(relatedMapper, {
         where: {
           [relatedMapper.idAttribute]: {
-            'in': unique(localKeys).filter((x) => x)
+            in: unique(localKeys).filter((x) => x)
           }
         }
       }, __opts).then((relatedItems) => {
         records.forEach((item) => {
-          let attached = []
+          const attached = []
           let itemKeys = utils.get(item, def.localKeys) || []
           itemKeys = utils.isArray(itemKeys) ? itemKeys : Object.keys(itemKeys)
           relatedItems.forEach((relatedItem) => {
@@ -1061,7 +1081,7 @@ Component.extend({
         return relatedItems
       })
     }
-  },
+  }
 
   loadHasManyForeignKeys (mapper, def, records, __opts) {
     const relatedMapper = def.getRelation()
@@ -1076,7 +1096,7 @@ Component.extend({
       return this.findAll(def.getRelation(), {
         where: {
           [def.foreignKeys]: {
-            'contains': this.makeHasManyForeignKeys(mapper, def, record)
+            contains: this.makeHasManyForeignKeys(mapper, def, record)
           }
         }
       }, __opts).then((relatedItems) => {
@@ -1086,7 +1106,7 @@ Component.extend({
       return this.findAll(relatedMapper, {
         where: {
           [def.foreignKeys]: {
-            'isectNotEmpty': records.map((record) => this.makeHasManyForeignKeys(mapper, def, record))
+            isectNotEmpty: records.map((record) => this.makeHasManyForeignKeys(mapper, def, record))
           }
         }
       }, __opts).then((relatedItems) => {
@@ -1104,7 +1124,7 @@ Component.extend({
         })
       })
     }
-  },
+  }
 
   /**
    * Load a hasOne relationship.
@@ -1127,7 +1147,7 @@ Component.extend({
         }
       })
     })
-  },
+  }
 
   /**
    * Return the foreignKey from the given record for the provided relationship.
@@ -1143,7 +1163,7 @@ Component.extend({
    */
   makeHasManyForeignKey (mapper, def, record) {
     return def.getForeignKey(record)
-  },
+  }
 
   /**
    * Return the localKeys from the given record for the provided relationship.
@@ -1160,7 +1180,7 @@ Component.extend({
     itemKeys = utils.isArray(itemKeys) ? itemKeys : Object.keys(itemKeys)
     localKeys = localKeys.concat(itemKeys)
     return unique(localKeys).filter((x) => x)
-  },
+  }
 
   /**
    * Return the foreignKeys from the given record for the provided relationship.
@@ -1173,7 +1193,7 @@ Component.extend({
    */
   makeHasManyForeignKeys (mapper, def, record) {
     return utils.get(record, mapper.idAttribute)
-  },
+  }
 
   /**
    * Return the foreignKey from the given record for the provided relationship.
@@ -1186,7 +1206,7 @@ Component.extend({
    */
   makeBelongsToForeignKey (mapper, def, record) {
     return def.getForeignKey(record)
-  },
+  }
 
   /**
    * Retrieve sum of the specified field of the records that match the selection
@@ -1218,7 +1238,7 @@ Component.extend({
 
     // beforeSum lifecycle hook
     op = opts.op = 'beforeSum'
-    return utils.resolve(this[op](mapper, field, query, opts))
+    return utils.resolve(this.beforeSum(mapper, field, query, opts))
       .then(() => {
         // Allow for re-assignment from lifecycle hook
         op = opts.op = 'sum'
@@ -1233,10 +1253,10 @@ Component.extend({
 
         // afterSum lifecycle hook
         op = opts.op = 'afterSum'
-        return utils.resolve(this[op](mapper, field, query, opts, response))
+        return utils.resolve(this.afterSum(mapper, field, query, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
 
   /**
    * @name Adapter#respond
@@ -1248,7 +1268,7 @@ Component.extend({
    */
   respond (response, opts) {
     return this.getOpt('raw', opts) ? response : response.data
-  },
+  }
 
   /**
    * Apply the given update to the record with the specified primary key. Called
@@ -1271,7 +1291,7 @@ Component.extend({
 
     // beforeUpdate lifecycle hook
     op = opts.op = 'beforeUpdate'
-    return utils.resolve(this[op](mapper, id, props, opts))
+    return utils.resolve(this.beforeUpdate(mapper, id, props, opts))
       .then((_props) => {
         // Allow for re-assignment from lifecycle hook
         props = _props === undefined ? props : _props
@@ -1289,10 +1309,10 @@ Component.extend({
 
         // afterUpdate lifecycle hook
         op = opts.op = 'afterUpdate'
-        return utils.resolve(this[op](mapper, id, props, opts, response))
+        return utils.resolve(this.afterUpdate(mapper, id, props, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
 
   /**
    * Apply the given update to all records that match the selection query.
@@ -1322,7 +1342,7 @@ Component.extend({
 
     // beforeUpdateAll lifecycle hook
     op = opts.op = 'beforeUpdateAll'
-    return utils.resolve(this[op](mapper, props, query, opts))
+    return utils.resolve(this.beforeUpdateAll(mapper, props, query, opts))
       .then((_props) => {
         // Allow for re-assignment from lifecycle hook
         props = _props === undefined ? props : _props
@@ -1341,10 +1361,10 @@ Component.extend({
 
         // afterUpdateAll lifecycle hook
         op = opts.op = 'afterUpdateAll'
-        return utils.resolve(this[op](mapper, props, query, opts, response))
+        return utils.resolve(this.afterUpdateAll(mapper, props, query, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
-  },
+  }
 
   /**
    * Update the given records in a single batch. Called by `Mapper#updateMany`.
@@ -1368,7 +1388,7 @@ Component.extend({
 
     // beforeUpdateMany lifecycle hook
     op = opts.op = 'beforeUpdateMany'
-    return utils.resolve(this[op](mapper, records, opts))
+    return utils.resolve(this.beforeUpdateMany(mapper, records, opts))
       .then((_records) => {
         // Allow for re-assignment from lifecycle hook
         records = _records === undefined ? records : _records
@@ -1387,61 +1407,8 @@ Component.extend({
 
         // afterUpdateMany lifecycle hook
         op = opts.op = 'afterUpdateMany'
-        return utils.resolve(this[op](mapper, records, opts, response))
+        return utils.resolve(this.afterUpdateMany(mapper, records, opts, response))
           .then((_response) => _response === undefined ? response : _response)
       })
   }
-})
-
-/**
- * Create a subclass of this Adapter:
- *
- * @example <caption>Adapter.extend</caption>
- * // Normally you would do: import {Adapter} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.10')
- * const {Adapter} = JSData
- * console.log('Using JSData v' + JSData.version.full)
- *
- * // Extend the class using ES2015 class syntax.
- * class CustomAdapterClass extends Adapter {
- *   foo () { return 'bar' }
- *   static beep () { return 'boop' }
- * }
- * const customAdapter = new CustomAdapterClass()
- * console.log(customAdapter.foo())
- * console.log(CustomAdapterClass.beep())
- *
- * // Extend the class using alternate method.
- * const OtherAdapterClass = Adapter.extend({
- *   foo () { return 'bar' }
- * }, {
- *   beep () { return 'boop' }
- * })
- * const otherAdapter = new OtherAdapterClass()
- * console.log(otherAdapter.foo())
- * console.log(OtherAdapterClass.beep())
- *
- * // Extend the class, providing a custom constructor.
- * function AnotherAdapterClass () {
- *   Adapter.call(this)
- *   this.created_at = new Date().getTime()
- * }
- * Adapter.extend({
- *   constructor: AnotherAdapterClass,
- *   foo () { return 'bar' }
- * }, {
- *   beep () { return 'boop' }
- * })
- * const anotherAdapter = new AnotherAdapterClass()
- * console.log(anotherAdapter.created_at)
- * console.log(anotherAdapter.foo())
- * console.log(AnotherAdapterClass.beep())
- *
- * @method Adapter.extend
- * @param {Object} [props={}] Properties to add to the prototype of the
- * subclass.
- * @param {Object} [props.constructor] Provide a custom constructor function
- * to be used as the subclass itself.
- * @param {Object} [classProps={}] Static properties to add to the subclass.
- * @returns {Constructor} Subclass of this Adapter class.
- */
+}
